// src/generated/api/models/PostListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostListResponse
 */
export interface PostListResponse {
    /**
     * 
     * @type {number}
     * @memberof PostListResponse
     */
    postId?: number;
    /**
     * 
     * @type {string}
     * @memberof PostListResponse
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PostListResponse
     */
    contentSnippet?: string;
    /**
     * 
     * @type {string}
     * @memberof PostListResponse
     */
    writer?: string;
    /**
     * 
     * @type {string}
     * @memberof PostListResponse
     */
    writerProfileImageUrl?: string;
    /**
     * 
     * @type {Date}
     * @memberof PostListResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof PostListResponse
     */
    tags?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof PostListResponse
     */
    thumbnailImageUrl?: string;
}

/**
 * Check if a given object implements the PostListResponse interface.
 */
export function instanceOfPostListResponse(value: object): value is PostListResponse {
    return true;
}

export function PostListResponseFromJSON(json: any): PostListResponse {
    return PostListResponseFromJSONTyped(json, false);
}

export function PostListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostListResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'postId': json['postId'] == null ? undefined : json['postId'],
        'title': json['title'] == null ? undefined : json['title'],
        'contentSnippet': json['contentSnippet'] == null ? undefined : json['contentSnippet'],
        'writer': json['writer'] == null ? undefined : json['writer'],
        'writerProfileImageUrl': json['writerProfileImageUrl'] == null ? undefined : json['writerProfileImageUrl'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'tags': json['tags'] == null ? undefined : json['tags'],
        'thumbnailImageUrl': json['thumbnailImageUrl'] == null ? undefined : json['thumbnailImageUrl'],
    };
}

export function PostListResponseToJSON(json: any): PostListResponse {
    return PostListResponseToJSONTyped(json, false);
}

export function PostListResponseToJSONTyped(value?: PostListResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'postId': value['postId'],
        'title': value['title'],
        'contentSnippet': value['contentSnippet'],
        'writer': value['writer'],
        'writerProfileImageUrl': value['writerProfileImageUrl'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'tags': value['tags'],
        'thumbnailImageUrl': value['thumbnailImageUrl'],
    };
}